Add BMS controls

+++ components/jk_bms/switch/jk_switch.h
@@ -11,7 +11,7 @@ class JkBms;
 class JkSwitch : public switch_::Switch, public Component {
  public:
   void set_parent(JkBms *parent) { this->parent_ = parent; };
-  void set_holding_register(uint16_t holding_register) { this->holding_register_ = holding_register; };
+  void set_holding_register(uint8_t holding_register) { this->holding_register_ = holding_register; };
   void dump_config() override;
   void loop() override {}
   float get_setup_priority() const override { return setup_priority::DATA; }
@@ -19,7 +19,7 @@ class JkSwitch : public switch_::Switch, public Component {
  protected:
   void write_state(bool state) override;
   JkBms *parent_;
-  uint16_t holding_register_;
+  uint8_t holding_register_;
 };
 
 }  // namespace jk_bms
